Modify App.js
1. App.js : events Module화 작업 진행.
