+++ mainApp/src/App.js
@@ -13,6 +13,11 @@ import PlayController from './components/playControllerComponent/PlayController'
 import utility from './utility/utility';
 import moment from 'moment'
 
+import playListEvents from "./events/playListEvents";
+import albumListEvents from "./events/albumListEvents";
+
+import searchListEvents from "./events/searchListEvents";
+import playControllerEvents from "./events/playControllerEvents";
 
 const ACTION_CONFIG = {
     addPlayList : "addPlayList",
@@ -23,7 +28,6 @@ const ACTION_CONFIG = {
     addAlbum : "addAlbum"
 };
 
-
 class App extends Component {
     constructor(props){
         super(props);
@@ -31,18 +35,29 @@ class App extends Component {
 
             //albumList
             albumList : null,
+
+            //선택된 현재 앨범
             currentAlbum: null,
             isAddClicked : false,
-            currentAlbumKey: 0,
 
 
             //playList,
+
+            //삭제 하는  video id
             deleteVideoCheckList : [],
+            //삭제 하는  video idx
             checkIdxList : [],
+            // 삭제 시 선택 하는 것
             selectAllIsChecked : false,
+            // play되는 player
             player: null,
+            //선택된 video 데이터
             selectedData : null,
+            //선택된 video index
             selectedKey : -1,
+
+
+            //현재 play되고 있는 video 데이터 (video, key, )
             playingState : null,
 
 
@@ -85,25 +100,27 @@ class App extends Component {
 
 
 
-        this.checkClickHandler = this.checkClickHandler.bind(this);
-        this.selectAllBtnClickHandler = this.selectAllBtnClickHandler.bind(this);
+
         this._getAlbumReqListener = this._getAlbumReqListener.bind(this);
-        this.deletePlayListBtnClickHandler = this.deletePlayListBtnClickHandler.bind(this);
         this._deletePlayListReqListener = this._deletePlayListReqListener.bind(this);
         this.onReady = this.onReady.bind(this);
 
 
         //playList
-        this.playListClickHandler = this.playListClickHandler.bind(this);
+        playListEvents.playListClickHandler = playListEvents.playListClickHandler.bind(this);
+        playListEvents.checkClickHandler = playListEvents.checkClickHandler.bind(this);
+        playListEvents.selectAllBtnClickHandler = playListEvents.selectAllBtnClickHandler.bind(this);
+        playListEvents.deletePlayListBtnClickHandler = playListEvents.deletePlayListBtnClickHandler.bind(this);
+
 
        //albumList
         this._getAllAlbumReqListener = this._getAllAlbumReqListener.bind(this);
-        this.albumClickHandler = this.albumClickHandler.bind(this);
-        this.deleteAlbumClickHandler = this.deleteAlbumClickHandler.bind(this);
+        albumListEvents.albumClickHandler = albumListEvents.albumClickHandler.bind(this);
+        albumListEvents.deleteAlbumClickHandler = albumListEvents.deleteAlbumClickHandler.bind(this);
         this._albumReqListener = this._albumReqListener.bind(this);
-        this.addAlbumSubmitHandler = this.addAlbumSubmitHandler.bind(this);
-        this.addItemClickHandler = this.addItemClickHandler.bind(this);
-        this.addItemCancelClickHandler = this.addItemCancelClickHandler.bind(this);
+        albumListEvents.addAlbumSubmitHandler = albumListEvents.addAlbumSubmitHandler.bind(this);
+        albumListEvents.addItemClickHandler = albumListEvents.addItemClickHandler.bind(this);
+        albumListEvents.addItemCancelClickHandler = albumListEvents.addItemCancelClickHandler.bind(this);
 
 
 
@@ -113,20 +130,18 @@ class App extends Component {
         this.videoArr = [];
         this.nextPageToken = "";
 
-        this.searchVideo = this.searchVideo.bind(this);
-        this.addSelectedVideo = this.addSelectedVideo.bind(this);
-        this.delSelectedVideo = this.delSelectedVideo.bind(this);
-        this.changeIsAllClearAddBtn = this.changeIsAllClearAddBtn.bind(this);
-        this.addSelectedVideoToAlbum = this.addSelectedVideoToAlbum.bind(this);
-        this.moreVideoList = this.moreVideoList.bind(this);
+        searchListEvents.searchVideo = searchListEvents.searchVideo.bind(this);
+        searchListEvents.addSelectedVideo = searchListEvents.addSelectedVideo.bind(this);
+        searchListEvents.delSelectedVideo = searchListEvents.delSelectedVideo.bind(this);
+        searchListEvents.changeIsAllClearAddBtn = searchListEvents.changeIsAllClearAddBtn.bind(this);
+        searchListEvents.addSelectedVideoToAlbum = searchListEvents.addSelectedVideoToAlbum.bind(this);
+        searchListEvents.moreVideoList = searchListEvents.moreVideoList.bind(this);
+        searchListEvents.initSearchList = searchListEvents.initSearchList.bind(this);
 
-        this.getVideoInfo=this.getVideoInfo.bind(this);
-        this.promiseSearch=this.promiseSearch.bind(this);
-        this.promiseGetViewCount=this.promiseGetViewCount.bind(this);
-        this.promiseGetDuration=this.promiseGetDuration.bind(this);
-
-
-        this.initSearchList = this.initSearchList.bind(this);
+        this._getVideoInfo=this._getVideoInfo.bind(this);
+        this._promiseSearch=this._promiseSearch.bind(this);
+        this._promiseGetViewCount=this._promiseGetViewCount.bind(this);
+        this._promiseGetDuration=this._promiseGetDuration.bind(this);
 
 
         //nav
@@ -134,32 +149,24 @@ class App extends Component {
 
 
         //playController
-        this.opts = {
-            videoId: this.state.videoId,
-            playerVars: { // https://developers.google.com/youtube/player_parameters
-                autoplay: 0
-            }
-        };
 
 
         this.interverId = null;
-        this.onChangePrevVideo = this.onChangePrevVideo.bind(this);
-        this.onChangeNextVideo = this.onChangeNextVideo.bind(this);
-        this.onPlayVideo = this.onPlayVideo.bind(this);
-        this.onPauseVideo = this.onPauseVideo.bind(this);
+        playControllerEvents.onChangePrevVideo = playControllerEvents.onChangePrevVideo.bind(this);
+        playControllerEvents.onChangeNextVideo = playControllerEvents.onChangeNextVideo.bind(this);
+        playControllerEvents.onPlayVideo = playControllerEvents.onPlayVideo.bind(this);
+        playControllerEvents.onPauseVideo = playControllerEvents.onPauseVideo.bind(this);
         this._setCurrentTime = this._setCurrentTime.bind(this);
         this.onPlayerStateChange = this.onPlayerStateChange.bind(this);
         this._setDuration = this._setDuration.bind(this);
-        this.moveSeekBar = this.moveSeekBar.bind(this);
+        playControllerEvents.moveSeekBar = playControllerEvents.moveSeekBar.bind(this);
 
         //sound
-        this.moveVolumeBar = this.moveVolumeBar.bind(this);
-        this.onSound = this.onSound.bind(this);
-        this.offSound = this.offSound.bind(this);
+        playControllerEvents.moveVolumeBar = playControllerEvents.moveVolumeBar.bind(this);
+        playControllerEvents.onSound = playControllerEvents.onSound.bind(this);
+        playControllerEvents.offSound = playControllerEvents.offSound.bind(this);
 
     }
-
-
 
 
     componentDidMount(){
@@ -167,44 +174,8 @@ class App extends Component {
     }
 
     //albumList
-    addItemClickHandler(event){
-        this.setState(()=>{
-               return {isAddClicked: true}
-            });
-        event.stopPropagation();
-
-    }
-    addItemCancelClickHandler(event){
-        this.setState(()=>{
-            return {isAddClicked: false}
-        });
-        event.stopPropagation();
-    }
 
 
-
-    addAlbumSubmitHandler(data,event){
-
-
-        let formData = new FormData();
-        //FormData 에 파일과 이메일을 append 메소드를 통해 등록
-
-        for(let key in data){
-            let inputData = data[key];
-            if(key === "category"){
-                inputData = JSON.stringify(inputData);
-            }
-            formData.append(key, inputData);
-        }
-
-        utility.runAjaxData(this._albumReqListener.bind(null,ACTION_CONFIG.addAlbum),"post","/albumList/addAlbum",formData);
-
-    }
-
-    deleteAlbumClickHandler(albumId,event){
-        utility.runAjax(this._albumReqListener.bind(null,ACTION_CONFIG.deleteAlbum), "GET", "/albumList/deleteAlbum/"+albumId);
-        event.stopPropagation();
-    }
 
     _albumReqListener(action,res){
         console.log(action);
@@ -218,51 +189,53 @@ class App extends Component {
        //console.log("jsonAlbumList",jsonAlbumList)
 
         switch (action){
+
+            // album 삭제 시
             case ACTION_CONFIG.deleteAlbum: this.setState((state)=>{
                 //console.log(jsonAlbumList);
-                if(!jsonAlbumList.err){
-                    let { playingState, currentAlbum } = state;
-                    let { playingAlbum,playingAlbumKey,playingData, playingKey } = playingState;
-                    if(playingState){
-                        if(currentAlbum !== playingAlbum) {
-                            return {
-                                albumList : jsonAlbumList,
-                                currentAlbum: jsonAlbumList[playingAlbumKey],
-                                currentAlbumKey : playingAlbumKey,
-                                selectedData : playingData,
-                                selectedKey : playingKey,
-                                deleteVideoCheckList : [],
-                                checkIdxList : [],
-                                selectAllIsChecked : false,
-                            }
-                        }
-
-                    }
-                    return {
-                        albumList : jsonAlbumList,
-                        currentAlbum: jsonAlbumList[0],
-                        currentAlbumKey : 0,
-                        deleteVideoCheckList : [],
-                        checkIdxList : [],
-                        selectAllIsChecked : false,
-                        player: null,
-                        selectedData : null,
-                        selectedKey : -1,
-                        playingState : null,
-                    }
-                }
-                return {
-                    albumList : null,
-                    currentAlbum: null,
-                    currentAlbumKey : 0,
+                // 엘범 리스크가 존재 하는 경우
+                let newState = {
                     deleteVideoCheckList : [],
                     checkIdxList : [],
                     selectAllIsChecked : false,
-                    player: null,
+                };
+                if(!jsonAlbumList.err){
+                    let { playingState, currentAlbum } = state;
+                    let { playingAlbum,playingData, playingKey } = playingState;
+
+                    //현재 play되고 있는 album이 존재하고 선택된 album과 play 되고 있는 album이 다를 경우
+                    if(playingState  && (currentAlbum._id !== playingAlbum._id)){
+                            return Object.assign({}, newState, {
+                                albumList : jsonAlbumList,
+                                currentAlbum: playingAlbum,
+                                selectedData : playingData,
+                                selectedKey : playingKey,
+                            });
+
+                    }
+
+                    return Object.assign({}, newState, {
+                        albumList : jsonAlbumList,
+                        currentAlbum: jsonAlbumList[0],
+                        selectedData : null,
+                        selectedKey : -1,
+                        player: null,
+                        playingState : null,
+                    });
+
+                }
+
+
+                // 마지막 앨범을 삭제 한 경우
+                return Object.assign({}, newState, {
+                    albumList : null,
+                    currentAlbum: null,
                     selectedData : null,
                     selectedKey : -1,
+                    player: null,
                     playingState : null,
-                }
+                });
+
 
             });
             break;
@@ -272,7 +245,6 @@ class App extends Component {
 
                     if(currentAlbum){
                         return {
-                            albumList : jsonAlbumList,
                             currentAlbum: currentAlbum,
                         }
                     }
@@ -288,16 +260,15 @@ class App extends Component {
             break;
             case ACTION_CONFIG.addAlbum: this.setState((state)=>{
                 if(!jsonAlbumList.err){
+                    // 최근에 추가된 album을 선택 해야 함
                     let currentAlbumKey = jsonAlbumList.length-1;
                     return {
                         albumList : jsonAlbumList,
                         currentAlbum: jsonAlbumList[currentAlbumKey],
-                        currentAlbumKey : currentAlbumKey,
                         isAddClicked : false,
                     }
                 }
             });
-
             break;
 
             default: break;
@@ -308,13 +279,7 @@ class App extends Component {
     }
 
 
-    albumClickHandler(_id,idx,event){
-        utility.runAjax(this._getAlbumReqListener.bind(null,ACTION_CONFIG.resetPlayList), "GET", "/albumList/getAlbum/"+_id);
-    }
-
     _getAlbumReqListener(action,res){
-        //
-        // console.log("jsonData",res.currentTarget.responseText);
 
         let jsonData = JSON.parse(res.currentTarget.responseText);
 
@@ -323,70 +288,60 @@ class App extends Component {
                 selectedVideoArr: [],
                 isSelectedArr: false,
                 isAllClearAddBtn: true,
-
                 totalDuration:0,
                 currentAlbum: jsonData
             });
             break;
+
             case ACTION_CONFIG.deletePlayList  : this.setState((state)=>{
 
                 let currentAlbum = jsonData;
-                let { playingState, checkIdxList, selectAllIsChecked,currentAlbumKey}  = state;
+                let { playingState, checkIdxList, selectAllIsChecked}  = state;
                 //어떤 동영상이 플레이 되고 있으면서 전체 선택이 안된 경우
+                let newState = {
+                    deleteVideoCheckList: [],
+                    checkIdxList: [],
+                    selectAllIsChecked: false,
+                    currentAlbum: jsonData
+                };
+
+
                 if(playingState && !selectAllIsChecked){
                     let { playingKey } = playingState;
                     let { playList } = currentAlbum;
                     //삭제 리스트에 있는 경우
                     if(checkIdxList.indexOf(playingKey) !== -1){
-
-                        return {
-
+                        return Object.assign({}, newState, {
                             playingState: Object.assign({}, playingState, {
-                                playingAlbum: currentAlbum,
+                                playingAlbum: jsonData,
                                 playingData: playList[0],
                                 playingKey: 0,
-                                playingAlbumKey : currentAlbumKey
                             }),
                             selectedData : playList[0],
                             selectedKey : 0,
-                            deleteVideoCheckList: [],
-                            checkIdxList: [],
-                            selectAllIsChecked: false,
-                            currentAlbum: jsonData
-                        }
+                        });
                     }
 
                     let length = checkIdxList.filter((val)=>{
                         return val < playingKey;
                     }).length;
                      let currentPlayingKey = playingKey-length;
-                      return {
-
+                      return Object.assign({}, newState, {
                           playingState: Object.assign({}, playingState, {
                               playingAlbum: currentAlbum,
                               playingData: playList[currentPlayingKey],
                               playingKey: currentPlayingKey,
-                              playingAlbumKey : currentAlbumKey
                           }),
                           selectedData : playList[currentPlayingKey],
                           selectedKey : currentPlayingKey,
-                          deleteVideoCheckList: [],
-                          checkIdxList: [],
-                          selectAllIsChecked: false,
-                          currentAlbum: jsonData
-                    }
+                      });
 
                 }
-                return {
+                return Object.assign({}, newState, {
                     playingState : null,
                     selectedData: null,
                     selectedKey: -1,
-                    deleteVideoCheckList: [],
-                    checkIdxList: [],
-                    selectAllIsChecked: false,
-                    currentAlbum: jsonData,
-                    // eventMap: Object.assign({},eventMap, resetEventMap),
-                };
+                });
 
 
             });
@@ -423,89 +378,12 @@ class App extends Component {
 
             });
         }
-
-
-
-    }
-
-
-
-    //삭제버튼 클릭 handler
-    deletePlayListBtnClickHandler(){
-        let { deleteVideoCheckList,currentAlbum } = this.state;
-        let { _id } = currentAlbum;
-        let deleteData = {
-            albumId : _id,
-            deleteList: deleteVideoCheckList
-        };
-        let jsonData = JSON.stringify(deleteData);
-        utility.runAjaxData(this._deletePlayListReqListener.bind(null,_id),"POST","/playList/deletePlayList", jsonData, "application/json");
     }
 
 
     _deletePlayListReqListener(_id,res){
         utility.runAjax(this._getAlbumReqListener.bind(null,ACTION_CONFIG.deletePlayList), "GET", "/albumList/getAlbum/"+_id);
     }
-
-    //check click handler
-    checkClickHandler(videoId, checkIdx, isChecked, event) {
-        let {deleteVideoCheckList,checkIdxList,currentAlbum}  = this.state;
-        //let items = videoData.items;
-        let currentSelectAllIsChecked = false;
-
-        let newDeleteVideoCheckList = [...deleteVideoCheckList];
-        let newCheckIdxList = [...checkIdxList];
-        //check 되어있지 않으면
-        if(!isChecked){
-            newDeleteVideoCheckList.push(videoId);
-            newCheckIdxList.push(checkIdx);
-        }else{
-            let idx =  checkIdxList.indexOf(checkIdx);
-            newCheckIdxList.splice(idx,1);
-            newDeleteVideoCheckList.splice(idx,1);
-        }
-
-
-        let playList = null;
-        if(currentAlbum){
-            playList = currentAlbum.playList;
-            if(newCheckIdxList.length === playList.length){
-                currentSelectAllIsChecked = true;
-            }
-        }
-
-        //console.log("currentSelectAllIsChecked",currentSelectAllIsChecked);
-
-        this.setState({deleteVideoCheckList: newDeleteVideoCheckList,checkIdxList:  newCheckIdxList, selectAllIsChecked: currentSelectAllIsChecked});
-        event.stopPropagation();
-    }
-
-    //전체선택 버튼 클릭 handler
-    selectAllBtnClickHandler(){
-        let { currentAlbum } = this.state;
-        let playList = null;
-
-        if(currentAlbum) {
-            playList = currentAlbum.playList;
-        }
-        let {selectAllIsChecked} = this.state;
-        let currentSelectAllIsChecked = false;
-        let newDeleteVideoCheckList = [];
-        let newCheckIdxList = [];
-
-
-        if(!selectAllIsChecked){
-            playList.forEach((val,idx)=>{
-                let {videoId} = val;
-                newDeleteVideoCheckList.push(videoId);
-                newCheckIdxList.push(idx);
-            });
-            currentSelectAllIsChecked = true;
-        }
-        this.setState({deleteVideoCheckList: newDeleteVideoCheckList,checkIdxList:  newCheckIdxList, selectAllIsChecked : currentSelectAllIsChecked});
-
-    }
-
 
 
 
@@ -522,129 +400,12 @@ class App extends Component {
     }
 
 
-    //playList
-    playListClickHandler(playList,key){
-
-                let {eventMap} = this.state;
-
-                this.setState((state) => {
-                    let {currentAlbum,currentAlbumKey, playingState} = state;
-                    let newEventMap = null;
-                    if(playingState) {
-                        let {playingKey} = playingState;
-                        if (key !== playingKey) {
-                            newEventMap = {
-                                playing: false,
-                                curTime: '00:00', // 현재 재생 시간
-                                totalTime: '00:00', // 전체 비디오 재생 시간
-                                curProgressBar: 0,
-                                maxProgressBar: 0,
-                            };
-                        }
-                    }
-
-                    return {
-                        playingState: Object.assign({}, playingState, {
-                            playingAlbum: currentAlbum,
-                            playingData: playList[key],
-                            playingKey: key,
-                            playingAlbumKey : currentAlbumKey
-                        }),
-                        selectedData: playList[key],
-                        selectedKey: key,
-                        eventMap: Object.assign({}, eventMap, newEventMap),
-
-                    };
-                }, ()=>{
-                    let {player} = this.state;
-                    if(player) {
-                        this._setDuration(player);
-                    }
-                });
-    }
 
 
 
     //searchList
 
-    addSelectedVideo(index){
-      let selectedItem = this.state.items[index];
-      let totalDuration = this.state.totalDuration + selectedItem.duration;
-
-        this.setState({
-            selectedVideoArr : this.state.selectedVideoArr.concat(selectedItem),
-            isSelectedArr : true,
-            totalDuration : totalDuration
-        })
-    }
-
-    delSelectedVideo(videoId){
-        let selectedVideoArr = [];
-        let isSelectedArr = false;
-
-        let totalDuration = 0;
-
-        selectedVideoArr = [...this.state.selectedVideoArr];
-
-        selectedVideoArr.forEach((data, index)=>{
-            if(data.videoId === videoId){
-              totalDuration = this.state.totalDuration - selectedVideoArr[index].duration;
-              selectedVideoArr.splice(index, 1);
-            }
-        });
-
-        if(selectedVideoArr.length !== 0){
-            isSelectedArr = true;
-        }
-
-        this.setState({
-            selectedVideoArr : [...selectedVideoArr],
-            isSelectedArr : isSelectedArr,
-
-            totalDuration : totalDuration
-        })
-    }
-
-    changeIsAllClearAddBtn(){
-        this.setState({
-            isAllClearAddBtn : false
-        })
-    }
-
-    addSelectedVideoToAlbum(_id){
-        let utilLayer = document.querySelector(".utilLayer");
-        utilLayer.classList.remove("show");
-
-        let insertData = {
-            albumId : _id,
-            selectedVideoArr : this.state.selectedVideoArr,
-            totalDuration : this.state.totalDuration
-        };
-        let jsonData = JSON.stringify(insertData);
-
-        utility.runAjaxData(function(e){
-            utility.runAjax(this._getAlbumReqListener.bind(null,ACTION_CONFIG.addPlayList), "GET", "/albumList/getAlbum/"+_id);
-        }.bind(this), "POST", "/playList/videos", jsonData, "application/json")
-    }
-
-
-    searchVideo(keyword){
-        this.setState({
-            items : [],
-            nextPageToken : "",
-            selectedVideoArr : [],
-            isSearched : true,
-            isSelectedArr : false
-
-        });
-        let encodedKeword = encodeURI(keyword);
-        this.searchUrl = "https://www.googleapis.com/youtube/v3/search?part=snippet&maxResults=15&q="+encodedKeword+"&key="+this.UTUBEKEY+"&type=video"
-
-        //this._searchAgainVideo(this.searchUrl);
-        this.getVideoInfo(this.searchUrl);
-    }
-
-    promiseSearch(url){
+    _promiseSearch(url){
       return new Promise(function(resolve, reject){
         utility.runAjax(function(e){
           let data = JSON.parse(e.target.responseText);
@@ -666,7 +427,7 @@ class App extends Component {
       }.bind(this))
     }
 
-    promiseGetViewCount(videoArr){
+    _promiseGetViewCount(videoArr){
       const UTUBEKEY = "AIzaSyDIkMgAKPVBeKhZcwdDo_ijqPiiK8DbYsA";
       let count = 0;
       return new Promise(function(resolve, reject){
@@ -689,7 +450,7 @@ class App extends Component {
       })
     }
 
-    promiseGetDuration(videoArr){
+    _promiseGetDuration(videoArr){
       const UTUBEKEY = "AIzaSyDIkMgAKPVBeKhZcwdDo_ijqPiiK8DbYsA";
       let count = 0;
       return new Promise(function(resolve, reject){
@@ -710,19 +471,18 @@ class App extends Component {
                 resolve(videoArr);
               }
             }
-
           }.bind(this), "GET", contentDetailsUrl)
         })
       })
     }
 
-    getVideoInfo(url){
-      this.promiseSearch(url)
+    _getVideoInfo(url){
+      this._promiseSearch(url)
         .then(function(videoArr){
-          return this.promiseGetViewCount(videoArr)
+          return this._promiseGetViewCount(videoArr)
         }.bind(this))
         .then(function(videoArr){
-          return this.promiseGetDuration(videoArr)
+          return this._promiseGetDuration(videoArr)
         }.bind(this))
         .then(function(videoArr){
           this.setState({
@@ -731,21 +491,6 @@ class App extends Component {
           })
         }.bind(this))
     }
-
-    moreVideoList(){
-        const url = this.searchUrl.concat("&pageToken="+this.state.nextPageToken);
-
-        let searchList = document.querySelector(".searchList");
-        let scrollHeight  = searchList.scrollHeight;
-        let clientHeight  = searchList.clientHeight;
-        let scrollTop  = searchList.scrollTop;
-
-        if((scrollHeight - scrollTop) === clientHeight){
-            //this._searchAgainVideo(url)
-            this.getVideoInfo(url)
-        }
-    }
-
 
     //navComponent
     navClickHandler(event){
@@ -764,149 +509,12 @@ class App extends Component {
 
     }
 
-    initSearchList(){
-      this.setState({
-        selectedVideoArr : [],
-        isSelectedArr : false,
-        isAllClearAddBtn : false,
-        totalDuration : 0
-      })
-    }
 
     //playController
     _toTimeString(seconds) {
         let time = (new Date(seconds * 1000)).toUTCString().match(/(\d\d:\d\d:\d\d)/)[0];
         time = time.replace(/00:/, "");
         return time
-    }
-
-
-    onPlayVideo(player) {
-        let {eventMap} = this.state;
-        let {playing} = eventMap;
-        if(!playing) {
-            this.setState((state)=>{
-                return {
-                    eventMap: Object.assign({}, eventMap, {
-                        playing: true ,
-                        totalTime: '00:00', // 전체 비디오 재생 시간
-                        maxProgressBar: 0,
-                    }),
-                }
-            },()=>{
-                player.playVideo();
-                this._setDuration(player);
-            });
-        }
-
-
-    }
-
-    onPauseVideo(player) {
-
-        event.stopPropagation();
-        event.preventDefault();
-        let {eventMap} = this.state;
-            if(eventMap.playing) {
-                this.setState((state) => {
-                    return {
-                        eventMap: Object.assign({}, eventMap, {playing: false}),
-                    }
-                },()=>{
-                    player.pauseVideo(); //  이게 state를 바꾸고 2번으로 바꾸는 작업 계속 발생한다
-                    this._setCurrentTime(player)
-                });
-            }
-
-
-
-    }
-
-    onChangeNextVideo() {
-        // this.setState({ videoId: selectedVideo.id.next });
-        let newEventMap = { playing: false,
-            curTime: '00:00', // 현재 재생 시간
-            totalTime: '00:00', // 전체 비디오 재생 시간
-            curProgressBar: 0,
-            maxProgressBar: 0,
-        };
-
-        let {eventMap} = this.state;
-
-        this.setState((state) => {
-            let { playingState, currentAlbum } = state;
-            let {playingAlbum, playingKey} = playingState;
-            let currentKey = playingKey+1;
-            let { playList } = playingAlbum;
-            if(currentKey > playList.length-1){
-                currentKey = 0;
-            }
-
-            let newSelectedData = null;
-            let newSelectedKey = -1;
-            if(playingAlbum === currentAlbum){
-                newSelectedData = playList[currentKey];
-                newSelectedKey = currentKey;
-            }
-
-            return {
-                playingState: Object.assign({}, playingState, {
-                    playingData: playList[currentKey],
-                    playingKey: currentKey,
-                }),
-                 selectedData: newSelectedData,
-                 selectedKey: newSelectedKey,
-                eventMap: Object.assign({}, eventMap, newEventMap),
-
-            };
-        },()=>{
-            let {player} = this.state;
-            if(player) {
-                this._setDuration(player);
-            }
-        });
-
-    }
-
-    onChangePrevVideo() {
-        let newEventMap = { playing: false,
-            curTime: '00:00', // 현재 재생 시간
-            totalTime: '00:00', // 전체 비디오 재생 시간
-            curProgressBar: 0,
-            maxProgressBar: 0,
-        };
-
-        let {eventMap} = this.state;
-        this.setState((state) => {
-            let { playingState,currentAlbum } = state;
-            let { playingAlbum, playingKey } = playingState;
-            let currentKey = playingKey-1;
-            let { playList } = playingAlbum;
-            if(currentKey < 0){
-                currentKey = playList.length-1;
-            }
-            let newSelectedData = null;
-            let newSelectedKey = -1;
-            if(playingAlbum === currentAlbum){
-                newSelectedData = playList[currentKey];
-                newSelectedKey = currentKey;
-            }
-            return {
-                playingState: Object.assign({}, playingState, {
-                    playingData: playList[currentKey],
-                    playingKey: currentKey,
-                }),
-                selectedData: newSelectedData,
-                selectedKey: newSelectedKey,
-                eventMap: Object.assign({}, eventMap, newEventMap),
-            };
-        },()=>{
-            let {player} = this.state;
-            if(player) {
-                this._setDuration(player);
-            }
-        });
-
     }
 
 
@@ -917,9 +525,9 @@ class App extends Component {
             this.setState({ videoState: event.data },()=>{
                 let { videoState } = this.state;
                 if(videoState===1){
-                    this.onPlayVideo(player);
+                    playControllerEvents.onPlayVideo(player);
                 }else if(videoState===2) {
-                    this.onPauseVideo(player);
+                    playControllerEvents.onPauseVideo(player);
                 }
                 else{
                     return;
@@ -982,7 +590,7 @@ class App extends Component {
                     }else if(curProgressBar >= maxProgressBar){
 
                         //끝나고 다음 행동을 여기서 정하면된다
-                        this.onChangeNextVideo();
+                        playControllerEvents.onChangeNextVideo();
 
 
                         /*
@@ -1000,67 +608,6 @@ class App extends Component {
             // console.log("fasdfasdfsadfasdf");
             clearInterval(this.interverId);
         }
-    }
-
-    moveSeekBar(player,event){
-        let bar = utility.$selector("#seekBar");
-        let maxVal = bar.max;
-        let curVal = bar.value;
-
-        player.seekTo(curVal, true);
-        if(maxVal === curVal){
-            this.onChangeNextVideo();
-        }
-
-    }
-
-
-    moveVolumeBar(player,event){
-        let bar = utility.$selector("#volumeBar");
-        let volumeVal = bar.value;
-        this.setState((state)=>{
-            let {eventMap}  = state;
-            return {
-                eventMap: Object.assign({}, eventMap, { volume: volumeVal ,preVolume: volumeVal}),
-            }
-
-        },()=>{
-            let {eventMap}  = this.state;
-            player.setVolume(volumeVal);
-            //console.log(volumeVal, this.state.event_map.volume);
-            if (eventMap.volume < 1){
-                this.offSound(player);
-            }
-            else {
-                this.onSound(player)
-            }
-        });
-    }
-
-    onSound(player){
-        this.setState((state)=> {
-            let {eventMap} = state;
-            let {preVolume} = eventMap;
-            return {
-                eventMap: Object.assign({}, eventMap, {soundOn: true, volume: preVolume})
-            };
-        },()=>{
-            player.unMute();
-        });
-
-    }
-
-    offSound(player){
-        this.setState((state)=> {
-                let {eventMap} = state;
-                return {
-                    eventMap: Object.assign({}, eventMap, {soundOn: false, volume: 0})
-                }
-            }
-            ,()=>{
-                player.mute();
-
-            });
     }
 
     render() {
@@ -1088,6 +635,7 @@ class App extends Component {
       let playList = null;
       let albumTitle = null;
       let _id = null;
+
       if(currentAlbum){
           _id = currentAlbum._id;
           albumTitle = currentAlbum.title;
@@ -1121,10 +669,10 @@ class App extends Component {
                 selectAllIsChecked={selectAllIsChecked}
                 onPlayerStateChange={this.onPlayerStateChange.bind(null,player)}
                 onReady={this.onReady}
-                playListClickHandler={this.playListClickHandler}
-                deletePlayListBtnClickHandler={this.deletePlayListBtnClickHandler}
-                selectAllBtnClickHandler={this.selectAllBtnClickHandler}
-                checkClickHandler={this.checkClickHandler}
+                playListClickHandler={playListEvents.playListClickHandler}
+                deletePlayListBtnClickHandler={playListEvents.deletePlayListBtnClickHandler}
+                selectAllBtnClickHandler={playListEvents.selectAllBtnClickHandler}
+                checkClickHandler={playListEvents.checkClickHandler}
             />
 
 
@@ -1135,24 +683,27 @@ class App extends Component {
                 //albumList
                 albumList={albumList}
                 isAddClicked={isAddClicked}
-                albumClickHandler={this.albumClickHandler}
-                deleteAlbumClickHandler={this.deleteAlbumClickHandler}
-                addAlbumSubmitHandler={this.addAlbumSubmitHandler}
-                addItemClickHandler={this.addItemClickHandler}
-                addItemCancelClickHandler={this.addItemCancelClickHandler}
+                albumClickHandler={albumListEvents.albumClickHandler}
+                deleteAlbumClickHandler={albumListEvents.deleteAlbumClickHandler}
+                addAlbumSubmitHandler={albumListEvents.addAlbumSubmitHandler}
+                addItemClickHandler={albumListEvents.addItemClickHandler}
+                addItemCancelClickHandler={albumListEvents.addItemCancelClickHandler}
 
                 //searchList
                 items={items}
                 isSelectedArr={isSelectedArr}
                 isAllClearAddBtn={isAllClearAddBtn}
                 isSearched = {isSearched}
-                addSelectedVideo={this.addSelectedVideo}
-                delSelectedVideo={this.delSelectedVideo}
-                changeIsAllClearAddBtn={this.changeIsAllClearAddBtn}
-                addSelectedVideoToAlbum={this.addSelectedVideoToAlbum.bind(null,_id)}
-                searchVideo={this.searchVideo}
-                moreVideoList={this.moreVideoList}
-                initSearchList = {this.initSearchList}
+
+
+
+                searchVideo={searchListEvents.searchVideo}
+                addSelectedVideo={searchListEvents.addSelectedVideo}
+                delSelectedVideo={searchListEvents.delSelectedVideo}
+                changeIsAllClearAddBtn={searchListEvents.changeIsAllClearAddBtn}
+                addSelectedVideoToAlbum={searchListEvents.addSelectedVideoToAlbum.bind(null,_id)}
+                moreVideoList={searchListEvents.moreVideoList}
+                initSearchList = {searchListEvents.initSearchList}
             />
             <Nav navClickHandler={this.navClickHandler}/>
 
@@ -1161,16 +712,16 @@ class App extends Component {
                 player={player}
                 eventMap={eventMap}
                 playingData={playingData}
-                onChangePrevVideo={this.onChangePrevVideo}
-                onChangeNextVideo={this.onChangeNextVideo}
-                onPlayVideo={this.onPlayVideo.bind(null,player)}
-                onPauseVideo={this.onPauseVideo.bind(null,player)}
+                onChangePrevVideo={playControllerEvents.onChangePrevVideo}
+                onChangeNextVideo={playControllerEvents.onChangeNextVideo}
+                onPlayVideo={playControllerEvents.onPlayVideo.bind(null,player)}
+                onPauseVideo={playControllerEvents.onPauseVideo.bind(null,player)}
                 onPlayerStateChange={this.onPlayerStateChange.bind(null,player)}
-                moveSeekBar={this.moveSeekBar.bind(null,player)}
+                moveSeekBar={playControllerEvents.moveSeekBar.bind(null,player)}
 
-                moveVolumeBar={this.moveVolumeBar.bind(null,player)}
-                onSound={this.onSound.bind(null,player)}
-                offSound={this.offSound.bind(null,player)}
+                moveVolumeBar={playControllerEvents.moveVolumeBar.bind(null,player)}
+                onSound={playControllerEvents.onSound.bind(null,player)}
+                offSound={playControllerEvents.offSound.bind(null,player)}
             />
       </div>
     );
